refactor: inline useMediaStreamer into InitialPlayer and merge LocalInitialPlayer
The useMediaStreamer hook and LocalInitialPlayer component were each
used by exactly one consumer. Inline both into InitialPlayer so the
component owns its streaming and local-file logic directly. This also
removes the nosleep.js dependency (replaced by Wake Lock API).
